Basic Damage System Implemented
Made a universal damage system component for enemies and players alike and simultaneously reworking the framework to easily allow it to function. To see it in action, press "J" to take away one health and "H" to see current health for the player. Also, added the player and enemy sprites to be used for the mvp